requirements.txt
asgiref==3.4.1
Django==3.2.5
django-seed==0.2.2
djangorestframework==3.12.4
Faker==8.10.1
Pillow==8.3.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
